post

+++ app/Http/Controllers/TasksController.php
@@ -8,24 +8,37 @@ use App\Task;    // 追加
 
 class TasksController extends Controller
 {
-    // getでmessages/にアクセスされた場合の「一覧表示処理」
+    // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
-    }
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            
+            //$tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            // タスク一覧を取得
+            $tasks = Task::all();
+            // タスク一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+            
+            $data = [
+                'user' => $user,
+                'tasklist' => $tasklist,
+            ];
+        }
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
+    } 
 
     // getでmessages/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
         $task = new Task;
 
-        // メッセージ作成ビューを表示
+        // タスク作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
         ]);
@@ -45,11 +58,13 @@ class TasksController extends Controller
             'content.required'=> '空文字を許さない',
         ]);
         
-        // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        // タスクを作成
+        //$request->user()->tasks()->create([
+            $task = new Task;
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        //]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -58,22 +73,22 @@ class TasksController extends Controller
     // getでmessages/（任意のid）にアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        
+        // タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
     }
-
     // getでmessages/（任意のid）/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
+        // タスク編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
         ]);
@@ -93,26 +108,35 @@ class TasksController extends Controller
             'content.required'=> '空文字を許さない',
         ]);
         
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、
+        if (\Auth::id() === $task->user_id) {
+            //タスクを編集
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
     }
 
-    // deleteでmessages/（任意のid）にアクセスされた場合の「削除処理」
+    // deleteで（任意のid）にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、
+        if (\Auth::id() === $task->user_id) {
+            //タスクを削除
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
+        
     }
 }
